Split AccessorEligibility registration helpers into dedicated classes
Keep the planner and cross-file predicates on AccessorEligibility, and move
site registration plus property-write shape rules so the file stays under the
length cap.
